# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ResearchQuestionInput {
  /**
     * @minLength 3
     * @maxLength 4000
     */
  question: string;
  /** @maxLength 120 */
  subject?: string;
  /** @maxLength 2000 */
  context?: string;
}

export interface ResearchResponse {
  answer: string;
  model: string;
}

export interface ApiError {
  error: string;
}

export interface BinderInsightInput {
  /**
     * @minLength 20
     * @maxLength 12000
     */
  binder: string;
  /**
     * @minLength 3
     * @maxLength 4000
     */
  update: string;
  todos: string[];
  completed: string[];
}

export interface BinderInsightResponse {
  completed: string[];
  add: string[];
  focus: string;
}

export interface BinderStructureInput {
  /**
     * @minLength 20
     * @maxLength 12000
     */
  binder: string;
}

export interface BinderSectionData {
  code: string;
  title: string;
  /** @minimum 1 */
  depth: number;
  body: string;
}

export interface BinderStructureResponse {
  sections: BinderSectionData[];
}