# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gemini
    description: Gemini-assisted research operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gemini/research:
    post:
      operationId: askGeminiResearch
      tags: [gemini]
      summary: Ask Gemini a research question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchQuestionInput"
      responses:
        "200":
          description: Research response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchResponse"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /gemini/binder-insights:
    post:
      operationId: updateGeminiBinderPlan
      tags: [gemini]
      summary: Turn binder progress into checklist updates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BinderInsightInput"
      responses:
        "200":
          description: Suggested binder plan updates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinderInsightResponse"
        "400":
          description: Invalid binder update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /gemini/binder-structure:
    post:
      operationId: analyzeBinderStructure
      tags: [gemini]
      summary: Parse a binder into its section structure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BinderStructureInput"
      responses:
        "200":
          description: Parsed binder structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinderStructureResponse"
        "400":
          description: Invalid binder text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ResearchQuestionInput:
      type: object
      properties:
        question:
          type: string
          minLength: 3
          maxLength: 4000
        subject:
          type: string
          maxLength: 120
        context:
          type: string
          maxLength: 2000
      required:
        - question
    ResearchResponse:
      type: object
      properties:
        answer:
          type: string
        model:
          type: string
      required:
        - answer
        - model
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BinderInsightInput:
      type: object
      properties:
        binder:
          type: string
          minLength: 20
          maxLength: 12000
        update:
          type: string
          minLength: 3
          maxLength: 4000
        todos:
          type: array
          items:
            type: string
        completed:
          type: array
          items:
            type: string
      required:
        - binder
        - update
        - todos
        - completed
    BinderInsightResponse:
      type: object
      properties:
        completed:
          type: array
          items:
            type: string
        add:
          type: array
          items:
            type: string
        focus:
          type: string
      required:
        - completed
        - add
        - focus
    BinderStructureInput:
      type: object
      properties:
        binder:
          type: string
          minLength: 20
          maxLength: 12000
      required:
        - binder
    BinderStructureResponse:
      type: object
      properties:
        sections:
          type: array
          items:
            $ref: "#/components/schemas/BinderSectionData"
      required:
        - sections
    BinderSectionData:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        depth:
          type: number
          minimum: 1
        body:
          type: string
      required:
        - code
        - title
        - depth
        - body