# lib/api-zod/src/generated/types/binderSectionData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BinderSectionData {
  code: string;
  title: string;
  /** @minimum 1 */
  depth: number;
  body: string;
}