# lib/api-zod/src/generated/types/binderStructureInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BinderStructureInput {
  /**
     * @minLength 20
     * @maxLength 12000
     */
  binder: string;
}