# lib/api-zod/src/generated/types/binderStructureResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BinderSectionData } from './binderSectionData';

export interface BinderStructureResponse {
  sections: BinderSectionData[];
}